docs: o2 specific
