+++ package.json
@@ -23,6 +23,10 @@
     "publish-beta": "npm run publish-beta:package && npm run publish-beta:extension",
     "publish-beta:package": "ts-node ./scripts/publish-beta-package.ts",
     "publish-beta:extension": "ts-node ./scripts/publish-beta-extension.ts",
+    "publish:o2": "npm run o2:general && npm run o2:build && npm run o2:upload",
+    "o2:general": "ts-node ./scripts/o2/general-pack",
+    "o2:build": "ts-node ./scripts/o2/build-pack",
+    "o2:release": "ts-node ./scripts/o2/release-pack",
     "lint": "eslint --cache --ext .ts,.tsx,.js,.jsx ./ --quiet",
     "lint:fix": "eslint --cache --ext .ts,.tsx,.js,.jsx ./ --quiet --fix",
     "coverage": "codecov",
