+++ scripts/o2/build-pack.ts
@@ -1,0 +1,16 @@
+async function installPackDeps() {
+  // TODO
+}
+
+async function packagePack() {
+  // TODO
+}
+
+async function buildPack() {
+  await installPackDeps();
+  await packagePack();
+}
+
+buildPack().catch((e) => {
+  console.error(e);
+});
