+++ scripts/o2/release-pack.ts
@@ -1,0 +1,7 @@
+async function uploadPack() {
+  // TODO
+}
+
+uploadPack().catch((e) => {
+  console.error(e);
+});
