+++ scripts/o2/template/index.ts.ejs
