+++ scripts/o2/template/package.json
