+++ scripts/o2/template/src/browser/index.ts
