+++ scripts/o2/template/src/index.ts
