+++ scripts/o2/template/tsconfig.json
